:mpl: Fix write_buffer; Remove bmp from wildcards list

+++ Lib/mwx/graphman.py
@@ -1471,7 +1471,6 @@ class Frame(mwx.Frame):
     ## --------------------------------
     wildcards = [
         "TIF file (*.tif)|*.tif",
-        "BMP file (*.bmp)|*.bmp",
          "ALL files (*.*)|*.*",
     ]
     
@@ -1494,11 +1493,12 @@ class Frame(mwx.Frame):
         """Write buffer to file path (to be overridden)."""
         try:
             img = Image.fromarray(buf)
-            img.save(path) # PIL saves as L,I,F,RGB.
+            img.save(path) # PIL saves as L, I, F, and RGB.
         except PermissionError:
             raise
-        except OSError: # e.g., cannot write mode L; 16 as BMP
-            os.remove(path)
+        except OSError: # cannot write mode L, I, F as BMP, etc.
+            if os.path.exists(path):
+                os.remove(path)
             raise
     
     def load_buffer(self, paths=None, view=None):
